ESP32_ADC_Characterization/requirements.txt
# Python dependencies for your project
# Add your project dependencies here, for example:
# requests==2.31.0
# pandas==2.1.0
# numpy==1.24.0

# Serial communication
pyserial==3.5

# To install dependencies, run:
# pip install -r requirements.txt
